Second version of gridlayout and viewlayout
- Now the grid item has 1 more field. Value field.
- Added first version of the "paytmlayout" based on paytm brand guide. Changes on actions bars and components colors were made.
- Value EditText is now properly formating the decimal value.
- Icon and Logo drawable added
- Color pallet defined on Color.xml

+++ app/src/main/java/com/example/bmoreira/paytmchallenge/MainMVP.java
@@ -19,7 +19,7 @@ public interface MainMVP {
         void onCreate();
         void onDestroy();
         void onBaseCurrencyChange(String currency);
-        void onAmountChange(int amount);
+        void onAmountChange(double amount);
     }
 
     interface Interactor {
